TVIST1-253: Made AbstractEntityListener class for logging and fixed namespace issue
