+++ config/services.yaml
@@ -38,6 +38,6 @@ services:
     # add more service definitions when explicit configuration is needed
     # please note that last definitions always *replace* previous ones
 
-    App\EntityListener\:
+    App\Logging\EntityListener\:
         resource: '../src/Logging/EntityListener/'
         tags: ['doctrine.orm.entity_listener']
